Fix: add missing meta files for unity libjpeg
Includes the missing meta file for the libjpeg wasm. This issue is difficult to catch locally because Unity generates the .meta files automatically for local packages, but not for those in the Package Cache.

Fortunately, with build automation as part of our CI soon, we'll catch this before it goes out since the missing .meta file prevents builds from completing.

Diffs=
932eb47d8 Fix: add missing meta files for unity libjpeg (#7384)

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com>