fix: revert min package version change
This PR temporarily reverts the min version change so that the last mega PR can be pushed downstream.

We have to do this because the github workflow that pushes the changes downstream doesn't love when we update the package json along with a bunch of other changes in one go, so we need to do the min version change separately.

Otherwise, we get this error: `Don't know what to do, applying diffs failed in an unexpected way`

Diffs=
8dd4a01bae fix: revert min package version change (#8834)
130c89c178 fix: conditionally compile components if ugui not present (#8829)

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com>