fix: revert min package version change
This PR temporarily reverts the min version change so that we can unblock unity changes from being pushed downstream.

We have to do this because the github workflow that pushes the changes downstream doesn't love when we update the package json along with a bunch of other changes in one go, so we need to do the min version change separately.

Otherwise, we get this error: `Don't know what to do, applying diffs failed in an unexpected way`

Diffs=
ff8013975f fix: revert min package version change (#8843)
6944998eba feat: update min unity version to 2021 (#8839)

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com>