fix: issue with rive panel not rendering on scene change
This PR addresses a bug in the Built-in Render Pipeline, where a RivePanel set to persist across scenes with DontDestroyOnLoad would stop rendering after a scene transition. This isn't a problem in URP and HDRP because they provide easy ways to start rendering when a camera is available.

To simulate this in the built-in RP, we detect when there's an available camera in the new scene and then use that as the render camera.

Diffs=
9fef272d63 fix: issue with rive panel not rendering on scene change (#8902)

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com>