fix: setting inputs on initial frame with RiveWidget
This PR fixes a reported issue where firing a trigger in `Start()` during the first frame would not register properly even though the `riveWidget.Status` is `Loaded`. This was happening because of a timing issue related to when rendering/advancing happens in relation to `Start()`.
Now we enforce using the `OnWidgetStateChange` event if you need to do anything related to the rive file on load. It still happens on the initial frame but at a safer time, and we no longer report that the widget is loaded until then.

Also addressed:

- Issue with the inspector potentially triggering an asset reload in the editor.
- Issue with default artboard/statemachine values when trying to load a new rive file into an existing RiveWidget.

Diffs=
6a92245f13 fix: setting inputs on initial frame with RiveWidget (#8914)

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com>