feat: get any camera in the scene if no main camera
This PR improves how the `Built-In RP Handler` and `URP Handler` select a camera to render the Rive Panel during scene transitions. Previously, it only attempted to fetch the `MainCamera`, but some use cases require assigning a different tag to the main camera, which could cause issues.

Now, if the `MainCamera` isn’t available, the handler will attempt to fetch other valid camera options in the scene.

Additionally, this PR exposes a static handle to the current `Built-In RP Handler`/ `URP Handler`, allowing users to manually set the `RenderCamera` if needed.

Diffs=
3cbd3b1e06 feat: get any camera in the scene if no main camera (#8929)

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com>