fix(vulkan): Fix image/gradient rendering on some Android devices (#10894) d8e7fbdb7b On some Adreno models, gradients were not properly rendering properly, due to the texture/gradient coordinates always being 0
This was caused by what appears to be a driver bug, in which doing vkUpdateDescriptorSets with multiple descriptors (in our case, updating paintData and paintAuxData in a single call) would fail to properly commit the data from the second one. We're only doing this in the one place. Separating this into two calls causes gradients and images to now display as expected.