Pushing merge, resolved using upstream.\n\n message=chore(scripting): add clipping support to scripted drawables (#11184) 27085dfb0c
fix(editor): Display runtime errors when Script input VM properties not set (#11177) cd70e30236
Fixes a crash when trying to use pointer events when a ViewModel property input is not setup properly on a scripted drawable instance.

The code was checking for ScriptAsset.vm which will always be non-null when bytecode is compiled in the editor. We should be checking instead for ScriptedObject.state because that will only be non-null if the ScriptedObject has validated (and we early out if that is null so we don't crash).

Also added some errors when VM properties are not setup correctly as Script inputs so users will be aware that there is an issue with their VM setup or bindings

chore: Refactor invalidateEffect for StrokeEffects (#11173) 05630acdb1
Refactors how invalidateEffect in StrokeEffects work. Now there are 2 code paths:

invalidateEffectsFromLocal() - this is called from the StrokeEffect itself. It calls out to ShapePaint::invalidateEffects to let it invalidate any other StrokeEffects that may need to do so. It also calls invalidateEffect on itself.
invalidateEffects() - this can be called from the StrokeEffect (via invalidateEffectsFromLocal) on itself OR by ShapePaint while in the process of determining if any other StrokeEffects should be invalidated via the call above. This adds paint dirt to the StrokeEffect along with any additional work needed when invalidating.
This prevents invalidateEffect from calling ShapePaint::invalidateEffects multiple times.

This also resolves an issue with ScriptedPathEffects in Components

fix(unity): align linear and gamma color space appearance (#11061) 62764bc262

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com> diff=\n\ndiff --cc package/package.json
index 128b0666,046af92c..00000000
--- a/package/package.json
+++ b/package/package.json
@@@ -1,21 -1,24 +1,46 @@@
  {
++<<<<<<< ours
 +    "name": "app.rive.rive-unity",
 +    "version": "0.3.9-canary.150",
 +    "displayName": "Rive",
 +    "description": "Create and ship interactive animations to any platform",
 +    "unity": "2020.1",
 +    "unityRelease": "0a10",
 +    "documentationUrl": "https://rive.app/community/doc/getting-started/docQycIiNOv4",
 +    "changelogUrl": "https://github.com/rive-app/rive-unity/tags",
 +    "licensesUrl": "https://github.com/rive-app/rive-unity/license",
 +    "keywords": [
 +        "Vector",
 +        "Graphics",
 +        "Animation"
 +    ],
 +    "author": {
 +        "name": "Rive Team",
 +        "email": "hello@rive.app",
 +        "url": "https://rive.app"
 +    }
++=======
+   "name": "app.rive.rive-unity",
+   "version": "0.0.0",
+   "displayName": "Rive",
+   "description": "Create and ship interactive animations to any platform",
+   "unity": "2021.3",
+   "unityRelease": "0a10",
+   "documentationUrl": "https://rive.app/community/doc/getting-started/docQycIiNOv4",
+   "changelogUrl": "https://github.com/rive-app/rive-unity/tags",
+   "licensesUrl": "https://github.com/rive-app/rive-unity/license",
+   "keywords": [
+     "Vector",
+     "Graphics",
+     "Animation"
+   ],
+   "author": {
+     "name": "Rive Team",
+     "email": "hello@rive.app",
+     "url": "https://rive.app"
+   },
+   "dependencies": {
+     "com.unity.shadergraph": "14.0.11"
+   }
++>>>>>>> theirs
  }